+++ test/shareMoney/shareMoney.t.sol
@@ -7,13 +7,15 @@ import "@hack/shareMoney/shareMoney.sol";
 
 contract ShareMoneyTest is Test {
     ShareMoney public sh;
+
     function setUp() public {
         sh = new ShareMoney();
     }
+
     function testReceive() public {
         address randAddress = vm.addr(1111);
         vm.startPrank(randAddress);
-        uint amount = 10000;
+        uint256 amount = 10000;
         vm.deal(randAddress, amount);
         uint256 balanceBefore = address(sh).balance;
         payable(address(sh)).transfer(10000);
@@ -22,12 +24,13 @@ contract ShareMoneyTest is Test {
         assertEq(address(randAddress).balance, 0, "the balanse need be 0");
         vm.stopPrank();
     }
+
     function testShare() external {
         payable(address(sh)).transfer(31);
         sh.share();
-        uint s = sh.part();
+        uint256 s = sh.part();
         assertEq(s, 1, "it need be 1");
-        uint user = address(sh.addressArray(10)).balance;
+        uint256 user = address(sh.addressArray(10)).balance;
         assertEq(user, 1, "error!!!");
     }
 }
