+++ test/wallet/wallet.t.sol
@@ -17,17 +17,13 @@ contract WalletTest is Test {
     function testDeposit() public {
         address randAdress = vm.addr(1111);
         vm.startPrank(randAdress);
-        uint amount = 50;
+        uint256 amount = 50;
         vm.deal(randAdress, amount);
         uint256 balanceBefore = address(w).balance;
         // Call the deposit function of the Wallet contract with 1 ether
         payable(address(w)).transfer(10);
         uint256 balanceAfter = address(w).balance;
-        assertEq(
-            balanceAfter - balanceBefore,
-            10,
-            "expect increase of 10 ether"
-        );
+        assertEq(balanceAfter - balanceBefore, 10, "expect increase of 10 ether");
         assertEq(address(randAdress).balance, 40, "abcde");
         vm.stopPrank();
     }
@@ -35,7 +31,7 @@ contract WalletTest is Test {
     function testWithdraw() public {
         address userAllow = 0xaC4E320Ed1235F185Bc6AC8856Ec7FEA7fF0310d;
         vm.startPrank(userAllow);
-        uint amount = 50;
+        uint256 amount = 50;
         vm.deal(address(w), amount);
         uint256 balanceBefore = address(userAllow).balance;
         w.withdraw(5);
@@ -48,7 +44,7 @@ contract WalletTest is Test {
     function testWithdrawNotAllow() public {
         address userNotAllow = 0x5ced660E3b925f034f99Df9466324F30A8Edf176;
         vm.startPrank(userNotAllow);
-        uint amount = 50;
+        uint256 amount = 50;
         vm.deal(address(w), amount);
         uint256 balanceBefore = address(userNotAllow).balance;
         vm.expectRevert();
